requirements.txt
gitpython
python-dotenv
langchain_openai 
langchain_core
langchain_community
faiss-cpu
streamlit
fastapi
uvicorn
sse_starlette